requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-google-genai
langchain-pinecone
pinecone-client
pypdf
python-docx
beautifulsoup4
numpy
pandas
requests
python-dotenv
aiofiles
sentence-transformers
tiktoken
google-generativeai
faiss-cpu
pydantic-settings
httpx
